Move `UserLocation` and `LocationCoordinates` to `$lib/types`
These types were defined in the `location` module but used elsewhere,
moving them to `$lib/types` prevents deep imports.

- Update imports
