+++ src/lib/components/LocationInput.svelte
@@ -2,7 +2,7 @@
 	import { getLocationService } from '$lib/location/context';
 	import { locationState } from '$lib/location/state.svelte';
 	import { createLogger, consoleProvider } from '$lib/logging';
-	import type { UserLocation } from '$lib/location';
+	import type { UserLocation } from '$lib/types';
 
 	const logger = createLogger(consoleProvider);
 	const service = getLocationService();
