+++ src/lib/location/index.ts
@@ -1,9 +1,3 @@
-export type {
-	LocationCoordinates,
-	UserLocation,
-	GeolocationProvider,
-	GeocodeProvider,
-	LocationService
-} from './types';
+export type { GeolocationProvider, GeocodeProvider, LocationService } from './types';
 
 export { locationState } from './state.svelte';
