+++ src/lib/types/index.ts
@@ -1,6 +1,5 @@
 // Interfaces and types for the app
 import type { SymptomName } from '$lib/config';
-import type { UserLocation } from '$lib/location';
 
 // User-readable pollen levels
 export type PollenRisk = 'low' | 'moderate' | 'high' | 'extreme';
@@ -37,4 +36,16 @@ export interface AppSettings {
 	id: number;
 	key: string;
 	value: unknown;
+}
+
+// Location Types
+export interface LocationCoordinates {
+	latitude: number;
+	longitude: number;
+}
+
+// Describes a single location
+export interface UserLocation {
+	label: string;
+	coordinates: LocationCoordinates;
 }
